requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
black==22.1.0
blue==0.9.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-ninja==1.3.0
django-phonenumber-field==8.0.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
email_validator==2.2.0
ffmpeg-python==0.2.0
flake8==4.0.1
future==1.0.0
idna==3.10
kombu==5.4.2
mccabe==0.6.1
mypy-extensions==1.0.0
pathspec==0.12.1
phonenumbers==8.13.51
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pycodestyle==2.8.0
pydantic==2.10.2
pydantic_core==2.27.1
pyflakes==2.4.0
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.2.0
requests==2.32.3
scalar-django-ninja @ git+https://github.com/scalar/scalar.git@c2f5f087b0e2494a3da0501c43c89bc876512d0c#subdirectory=packages/scalar_django_ninja
six==1.17.0
sqlparse==0.5.2
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
